chore(award) : move mapper class
