+++ src/main/java/org/helloworld/gymmate/domain/user/trainer/award/service/AwardService.java
@@ -2,6 +2,7 @@ package org.helloworld.gymmate.domain.user.trainer.award.service;
 
 import org.helloworld.gymmate.domain.user.trainer.award.dto.AwardRequest;
 import org.helloworld.gymmate.domain.user.trainer.award.entity.Award;
+import org.helloworld.gymmate.domain.user.trainer.award.mapper.AwardMapper;
 import org.helloworld.gymmate.domain.user.trainer.award.repository.AwardRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
